pyproject.toml
[tool.poetry]
name = "linz_logger"
version = "0.9.0"
description = "LINZ standard Logging format"
authors = ["Blayne <bchard@linz.govt.nz>", "Megan <mdavidson@linz.govt.nz>", "Victor <vengmark@linz.govt.nz>", "Alice <afage@linz.govt.nz>", "Paul <pfouquet@linz.govt.nz>"]
license = "MIT"

readme = "README.md"

repository = "https://github.com/linz/python-linz-logger"

keywords = ["LINZ", "logging"]

[tool.poetry.dependencies]
python = "^3.8"
structlog = "^22.1.0"
python-ulid = "^1.1.0"
dataclasses-json = "^0.5.7"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
pylint = "^2.15.0"
pytest = "^7.1.2"
pytest-dependency = "^0.5.1"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 127

[tool.isort]
line_length = 127
case_sensitive = true
use_parentheses = true
include_trailing_comma = true
multi_line_output = 3
force_grid_wrap = 0
